+++ src/storage.ts
@@ -89,7 +89,7 @@ export const localStore: PersistentStore = {
         }
 
         let supported = true
-        if (window) {
+        if (typeof window !== 'undefined') {
             try {
                 const key = '__mplssupport__',
                     val = 'xyz'
@@ -236,7 +236,7 @@ export const sessionStore: PersistentStore = {
             return sessionStorageSupported
         }
         sessionStorageSupported = true
-        if (window) {
+        if (typeof window !== 'undefined') {
             try {
                 const key = '__support__',
                     val = 'xyz'